Add create order endpoint

+++ src/main/java/com/bookstore/boundary/controller/product/CreateProductDocument.java
@@ -24,7 +24,7 @@ public class CreateProductDocument {
     @NotNull
     @PositiveOrZero
     @JsonProperty("inStock")
-    private final Double inStock;
+    private final Integer inStock;
 
     @NotNull
     @Positive
