+++ src/main/java/com/bookstore/domain/product/dto/CreateProductDto.java
@@ -13,7 +13,7 @@ public class CreateProductDto {
 
     private final String description;
 
-    private final Double inStock;
+    private final Integer inStock;
 
     private final Double price;
